requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
requests
oauth2client